+++ src/defined_gamut/mod.rs
@@ -10,7 +10,9 @@ use crate::Colour;
 
 /// Denotes a colour with a well-defined gamut.
 ///
-/// Types such as [`Rgb`](crate::rgb::Rgb) do not define this trait as they their gamuts are relative to the final display.
+/// More specifically, this trait denotes colours that have an "ideal" representation, i.e. each well-defined value can, in theory, only be displayed in one way.
+/// Examples include [`CieXyz`](crate::xyz::CieXyz) and [`OpRgb`](crate::rgb::OpRgb).
 ///
-/// Note that colours that define this trait may still be further transformed before being displayed.
+/// Types such as [`Rgb`](crate::rgb::Rgb) do not define this trait as their gamuts are unspecified.
+/// Note that colours that do define this trait may still be further transformed before being displayed.
 pub trait DefinedGamut: Colour { }