+++ src/error/html_from_str_error.rs
@@ -3,14 +3,17 @@
 use core::error::Error;
 use core::fmt::{self, Display, Formatter};
 
-/// A [HTML colour](crate::www::Html) could not be parsed.
+/// An [HTML colour](crate::www::Html) could not be parsed.
 #[derive(Debug, Eq, PartialEq)]
 pub enum HtmlFromStrError {
-	/// The HTML colour was missing a hash `#`.
+	/// An HTML hexadecimal colour was missing a hash `#`.
 	MissingHash,
 
-	/// The HTML colour had an otherwise unknown format.
+	/// An HTML colour had an otherwise unknown format.
 	UnknownFormat,
+
+	/// An HTML named colour was unknown.
+	UnknownName,
 }
 
 impl Error for HtmlFromStrError { }
@@ -20,10 +23,13 @@ impl Display for HtmlFromStrError {
 	fn fmt(&self, f: &mut Formatter) -> fmt::Result {
 		match *self {
 			Self::MissingHash
-			=> write!(f, "html colour is missing prefixed hash `#`"),
+			=> write!(f, "html hexadecimal colour is missing prefixed hash `#`"),
 
 			Self::UnknownFormat
 			=> write!(f, "html colour is of an otherwise unknown format"),
+
+			Self::UnknownName
+			=> write!(f, "html named colour is unknown"),
 		}
 	}
 }