+++ src/hsv/hwb/mod.rs
@@ -6,17 +6,25 @@
 // can obtain one at:
 // <https://mozilla.org/MPL/2.0/>.
 
-use crate::{Colour, Component};
-use crate::hsv::Hwba;
+use crate::{BalancedColour, Colour, Component};
+
+#[cfg(feature = "bytemuck")]
+use bytemuck::{Pod, Zeroable};
+
+#[cfg(feature = "serde")]
+use serde::{Deserialize, Serialize};
+
+#[cfg(feature = "zerocopy")]
+use zerocopy::{FromZeros, Immutable, IntoBytes};
 
 /// An HWB colour.
 ///
 /// This type guarantees that its three channels -- hue, whiteness, and blackness -- are stored sequentially in memory (in this order).
 #[repr(transparent)]
 #[derive(Clone, Copy, Debug, Default, Eq, Ord, PartialEq, PartialOrd)]
-#[cfg_attr(feature = "bytemuck", derive(bytemuck::Pod, bytemuck::Zeroable))]
-#[cfg_attr(feature = "serde",    derive(serde::Deserialize, serde::Serialize))]
-#[cfg_attr(feature = "zerocopy", derive(zerocopy::FromZeros, zerocopy::Immutable, zerocopy::IntoBytes))]
+#[cfg_attr(feature = "bytemuck", derive(Pod, Zeroable))]
+#[cfg_attr(feature = "serde",    derive(Deserialize, Serialize))]
+#[cfg_attr(feature = "zerocopy", derive(FromZeros, Immutable, IntoBytes))]
 pub struct Hwb<T>([T; 0x3]);
 
 impl<T: Component> Hwb<T> {
@@ -28,14 +36,6 @@ impl<T: Component> Hwb<T> {
 		Self(data)
 	}
 
-	/// Adds an alpha channel to the HWB colour.
-	#[inline(always)]
-	#[must_use]
-	pub const fn with_alpha(self, alpha: T) -> Hwba<T> {
-		let (hue, whiteness, blackness) = self.get();
-		Hwba::new(hue, whiteness, blackness, alpha)
-	}
-
 	/// Deconstructs an HWB colour.
 	#[inline(always)]
 	#[must_use]
@@ -43,6 +43,10 @@ impl<T: Component> Hwb<T> {
 		let [hue, whiteness, blackness] = self.0;
 		(hue, whiteness, blackness)
 	}
+}
+
+unsafe impl<T: Component> BalancedColour for Hwb<T> {
+	type Component = T;
 }
 
 impl<T: Component> Colour for Hwb<T> { }