+++ src/hsv/mod.rs
@@ -6,18 +6,12 @@
 // can obtain one at:
 // <https://mozilla.org/MPL/2.0/>.
 
-//! HSV- and HSL-based colour facilities.
+//! HSV-based colour facilities.
 
 mod hsl;
-mod hsla;
 mod hsv;
-mod hsva;
 mod hwb;
-mod hwba;
 
 pub use hsl::Hsl;
-pub use hsla::Hsla;
 pub use hsv::Hsv;
-pub use hsva::Hsva;
 pub use hwb::Hwb;
-pub use hwba::Hwba;