+++ src/lch/cie_lch/mod.rs
@@ -6,17 +6,25 @@
 // can obtain one at:
 // <https://mozilla.org/MPL/2.0/>.
 
-use crate::{Colour, Component, DefinedGamut};
-use crate::lch::CieLcha;
+use crate::{BalancedColour, Colour, Component, DefinedGamut};
+
+#[cfg(feature = "bytemuck")]
+use bytemuck::{Pod, Zeroable};
+
+#[cfg(feature = "serde")]
+use serde::{Deserialize, Serialize};
+
+#[cfg(feature = "zerocopy")]
+use zerocopy::{FromZeros, Immutable, IntoBytes};
 
 /// A CIELCh colour.
 ///
 /// This type guarantees that its three channels -- luminance, chroma, and hue -- are stored sequentially in memory (in this order).
 #[repr(transparent)]
 #[derive(Clone, Copy, Debug, Default, Eq, Ord, PartialEq, PartialOrd)]
-#[cfg_attr(feature = "bytemuck", derive(bytemuck::Pod, bytemuck::Zeroable))]
-#[cfg_attr(feature = "serde",    derive(serde::Deserialize, serde::Serialize))]
-#[cfg_attr(feature = "zerocopy", derive(zerocopy::FromZeros, zerocopy::Immutable, zerocopy::IntoBytes))]
+#[cfg_attr(feature = "bytemuck", derive(Pod, Zeroable))]
+#[cfg_attr(feature = "serde",    derive(Deserialize, Serialize))]
+#[cfg_attr(feature = "zerocopy", derive(FromZeros, Immutable, IntoBytes))]
 pub struct CieLch<T>([T; 0x3]);
 
 impl<T: Component> CieLch<T> {
@@ -28,14 +36,6 @@ impl<T: Component> CieLch<T> {
 		Self(data)
 	}
 
-	/// Adds an alpha channel to the CIELCh colour.
-	#[inline(always)]
-	#[must_use]
-	pub const fn with_alpha(self, alpha: T) -> CieLcha<T> {
-		let (x, y, z) = self.get();
-		CieLcha::new(x, y, z, alpha)
-	}
-
 	/// Deconstructs a CIELCh colour.
 	#[inline(always)]
 	#[must_use]
@@ -43,6 +43,10 @@ impl<T: Component> CieLch<T> {
 		let [x, y, z] = self.0;
 		(x, y, z)
 	}
+}
+
+unsafe impl<T: Component> BalancedColour for CieLch<T> {
+	type Component = T;
 }
 
 impl<T: Component> Colour for CieLch<T> { }