+++ src/rgb/op_rgb/mod.rs
@@ -6,18 +6,27 @@
 // can obtain one at:
 // <https://mozilla.org/MPL/2.0/>.
 
-use crate::{Colour, Component, DefinedGamut};
-use crate::rgb::{OpRgba, Rgb};
+use crate::{BalancedColour, Colour, Component, DefinedGamut};
+use crate::rgb::Rgb;
 
-/// An opRGB colour.
+#[cfg(feature = "bytemuck")]
+use bytemuck::{Pod, Zeroable};
+
+#[cfg(feature = "serde")]
+use serde::{Deserialize, Serialize};
+
+#[cfg(feature = "zerocopy")]
+use zerocopy::{FromZeros, Immutable, IntoBytes};
+
+/// An opRGB (also known as Adobe RGB) colour.
 ///
 /// This type is guaranteed to always have the exact same layout as <code>[Rgb]&lt;T&gt;</code>, although with a different representation.
 /// Namely, channel order is preserved in memory.
 #[repr(transparent)]
 #[derive(Clone, Copy, Debug, Default, Eq, Ord, PartialEq, PartialOrd)]
-#[cfg_attr(feature = "bytemuck", derive(bytemuck::Pod, bytemuck::Zeroable))]
-#[cfg_attr(feature = "serde",    derive(serde::Deserialize, serde::Serialize))]
-#[cfg_attr(feature = "zerocopy", derive(zerocopy::FromZeros, zerocopy::Immutable, zerocopy::IntoBytes))]
+#[cfg_attr(feature = "bytemuck", derive(Pod, Zeroable))]
+#[cfg_attr(feature = "serde",    derive(Deserialize, Serialize))]
+#[cfg_attr(feature = "zerocopy", derive(FromZeros, Immutable, IntoBytes))]
 pub struct OpRgb<T>(Rgb<T>);
 
 impl<T: Component> OpRgb<T> {
@@ -27,14 +36,6 @@ impl<T: Component> OpRgb<T> {
 	pub const fn new(red: T, green: T, blue: T) -> Self {
 		let colour = Rgb::new(red, green, blue);
 		Self(colour)
-	}
-
-	/// Adds an alpha channel to the opRGB colour.
-	#[inline(always)]
-	#[must_use]
-	pub const fn with_alpha(self, alpha: T) -> OpRgba<T> {
-		let (red, green, blue) = self.get();
-		OpRgba::new(red, green, blue, alpha)
 	}
 
 	/// Reinterprets the opRGB colour as a raw RGB colour.
@@ -50,6 +51,10 @@ impl<T: Component> OpRgb<T> {
 	pub const fn get(self) -> (T, T, T) {
 		self.as_rgb().get()
 	}
+}
+
+unsafe impl<T: Component> BalancedColour for OpRgb<T> {
+	type Component = T;
 }
 
 impl<T: Component> Colour for OpRgb<T> { }