+++ src/rgb/rgb/mod.rs
@@ -6,8 +6,16 @@
 // can obtain one at:
 // <https://mozilla.org/MPL/2.0/>.
 
-use crate::{Colour, Component};
-use crate::rgb::Rgba;
+use crate::{BalancedColour, Colour, Component};
+
+#[cfg(feature = "bytemuck")]
+use bytemuck::{Pod, Zeroable};
+
+#[cfg(feature = "serde")]
+use serde::{Deserialize, Serialize};
+
+#[cfg(feature = "zerocopy")]
+use zerocopy::{FromZeros, Immutable, IntoBytes};
 
 /// A raw RGB colour.
 ///
@@ -16,9 +24,9 @@ use crate::rgb::Rgba;
 /// Unlike other colours such as [SRgb](crate::rgb::SRgb), this one does not define a gamut *per se*; instead, users of this type are to interpret it on their own.
 #[repr(transparent)]
 #[derive(Clone, Copy, Debug, Default, Eq, Ord, PartialEq, PartialOrd)]
-#[cfg_attr(feature = "bytemuck", derive(bytemuck::Pod, bytemuck::Zeroable))]
-#[cfg_attr(feature = "serde",    derive(serde::Deserialize, serde::Serialize))]
-#[cfg_attr(feature = "zerocopy", derive(zerocopy::FromZeros, zerocopy::Immutable, zerocopy::IntoBytes))]
+#[cfg_attr(feature = "bytemuck", derive(Pod, Zeroable))]
+#[cfg_attr(feature = "serde",    derive(Deserialize, Serialize))]
+#[cfg_attr(feature = "zerocopy", derive(FromZeros, Immutable, IntoBytes))]
 pub struct Rgb<T>([T; 0x3]);
 
 impl<T: Component> Rgb<T> {
@@ -30,14 +38,6 @@ impl<T: Component> Rgb<T> {
 		Self(data)
 	}
 
-	/// Adds an alpha channel to the raw RGB colour.
-	#[inline(always)]
-	#[must_use]
-	pub const fn with_alpha(self, alpha: T) -> Rgba<T> {
-		let (red, green, blue) = self.get();
-		Rgba::new(red, green, blue, alpha)
-	}
-
 	/// Deconstructs a raw RGB colour.
 	#[inline(always)]
 	#[must_use]
@@ -45,6 +45,10 @@ impl<T: Component> Rgb<T> {
 		let [red, green, blue] = self.0;
 		(red, green, blue)
 	}
+}
+
+unsafe impl<T: Component> BalancedColour for Rgb<T> {
+	type Component = T;
 }
 
 impl<T: Component> Colour for Rgb<T> { }