+++ src/xyz/cie_xyz/mod.rs
@@ -7,16 +7,24 @@
 // <https://mozilla.org/MPL/2.0/>.
 
 use crate::{Colour, Component, DefinedGamut};
-use crate::xyz::CieXyza;
+
+#[cfg(feature = "bytemuck")]
+use bytemuck::{Pod, Zeroable};
+
+#[cfg(feature = "serde")]
+use serde::{Deserialize, Serialize};
+
+#[cfg(feature = "zerocopy")]
+use zerocopy::{FromZeros, Immutable, IntoBytes};
 
 /// A CIEXYZ colour.
 ///
 /// This type guarantees that its three channels -- X, Y, and Z -- are stored sequentially in memory (in this order).
 #[repr(transparent)]
 #[derive(Clone, Copy, Debug, Default, Eq, Ord, PartialEq, PartialOrd)]
-#[cfg_attr(feature = "bytemuck", derive(bytemuck::Pod, bytemuck::Zeroable))]
-#[cfg_attr(feature = "serde",    derive(serde::Deserialize, serde::Serialize))]
-#[cfg_attr(feature = "zerocopy", derive(zerocopy::FromZeros, zerocopy::Immutable, zerocopy::IntoBytes))]
+#[cfg_attr(feature = "bytemuck", derive(Pod, Zeroable))]
+#[cfg_attr(feature = "serde",    derive(Deserialize, Serialize))]
+#[cfg_attr(feature = "zerocopy", derive(FromZeros, Immutable, IntoBytes))]
 pub struct CieXyz<T>([T; 0x3]);
 
 impl<T: Component> CieXyz<T> {
@@ -26,14 +34,6 @@ impl<T: Component> CieXyz<T> {
 	pub const fn new(x: T, y: T, z: T) -> Self {
 		let data = [x, y, z];
 		Self(data)
-	}
-
-	/// Adds an alpha channel to the CIEXYZ colour.
-	#[inline(always)]
-	#[must_use]
-	pub const fn with_alpha(self, alpha: T) -> CieXyza<T> {
-		let (x, y, z) = self.get();
-		CieXyza::new(x, y, z, alpha)
 	}
 
 	/// Deconstructs a CIEXYZ colour.